refactor: 패키지 분류 및 파일 재배치
