+++ src/test/java/com/springboot/monew/notification/scheduler/NotificationCleanUpSchedulerTest.java
@@ -5,20 +5,16 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import com.springboot.monew.metric.ScheduledTaskMetrics;
-import com.springboot.monew.metric.notification.NotificationMetrics;
+import com.springboot.monew.common.metric.ScheduledTaskMetrics;
+import com.springboot.monew.notification.metric.NotificationMetrics;
 import com.springboot.monew.notification.service.NotificationService;
 import java.time.Clock;
 import java.time.Instant;
 import java.time.ZoneId;
-import org.instancio.Instancio;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
